requirements.txt
pyautogen
streamlit
yfinance
matplotlib